requirements.txt
requests
pytz
psycopg2-binary==2.9.9
